마이페이지 embed 구조로 인한 오류 해결 + @
1.
-결제 후 바로 구매 내역으로 이동하는 링크가 헤더 없는 페이지로 되어있는 문제 : 마이페이지에게 파라미터를 받게 하고 이에 따라 embed의 주소를 유동적으로 바꿀 수 있게 처리
-embed 영역 내에서 다른 링크로 이동하면 헤더가 2개가 되는 문제 : 부모창에서 작동하게 하는 함수 추가
2. 결제 페이지 예외 처리, 최근 본 상품 코드 살짝 개선, 디자인 살짝 교정
또 오류 발견되면 말씀 부탁드립니다.

+++ B/src/main/java/com/B/controller/IndexController.java
@@ -3,7 +3,6 @@ package com.B.controller;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
@@ -140,19 +139,16 @@ public class IndexController {
 			List<Map<String, Object>> recentLogList = mypageService.getRecentLogList(map.getMap());
 			int result = 0;
 			 for(int i = 0; i < recentLogList.size(); i++){
-				 for (int j = 0; j < recentLogList.get(i).size(); j++) {
-					 if(map.get("p_no") == recentLogList.get(i).get("p_no")) {
+					 if(map.get("p_no").toString().equals(recentLogList.get(i).get("p_no").toString())) {
 						 result = 1;
 					 }
-				}
 			  }
 			if(result == 0) mypageService.inputRecentLog(map.getMap());
 
 			//최근 본 상품은 10개까지 저장하도록 처리
 			result = mypageService.countRecentLog(map.getMap());
-			if (result > 10) {
-				mypageService.deleteRecentLog(map.getMap());
-			}
+			if (result > 10) mypageService.deleteRecentLog(map.getMap());
+
 		}
 		
 		return mv;
